add number answer adapter
