+++ app/src/main/java/org/stepik/android/adaptive/data/model/Reply.kt
@@ -3,5 +3,6 @@ package org.stepik.android.adaptive.data.model
 
 class Reply(
         private val choices: BooleanArray? = null,
-        private val text: String? = null
+        private val text: String? = null,
+        private var number: String? = null
 )
