+++ app/src/main/java/org/stepik/android/adaptive/resolvers/StepTypeResolverImpl.kt
@@ -1,12 +1,8 @@
 package org.stepik.android.adaptive.resolvers
 
 import org.stepik.android.adaptive.data.model.Step
-import org.stepik.android.adaptive.di.AppSingleton
 import org.stepik.android.adaptive.di.study.StudyScope
-import org.stepik.android.adaptive.ui.adapter.attempts.AttemptAnswerAdapter
-import org.stepik.android.adaptive.ui.adapter.attempts.ChoiceQuizAnswersAdapter
-import org.stepik.android.adaptive.ui.adapter.attempts.NotSupportedQuizAnswerAdapter
-import org.stepik.android.adaptive.ui.adapter.attempts.StringQuizAnswerAdapter
+import org.stepik.android.adaptive.ui.adapter.attempts.*
 import org.stepik.android.adaptive.util.AppConstants
 import javax.inject.Inject
 
@@ -18,8 +14,9 @@ constructor(): StepTypeResolver {
         val type = step.block?.name ?: return NotSupportedQuizAnswerAdapter()
 
         return when (type) {
-            AppConstants.TYPE_CHOICE -> ChoiceQuizAnswersAdapter()
+            AppConstants.TYPE_CHOICE -> ChoiceQuizAnswerAdapter()
             AppConstants.TYPE_STRING -> StringQuizAnswerAdapter()
+            AppConstants.TYPE_NUMBER -> NumberQuizAnswerAdapter()
 
             else -> NotSupportedQuizAnswerAdapter()
         }
